Update dependencies for Tasker
This update shouldn't affect how the program operates, but it does
update all the dependencies used by the program. The most notable
changes are that Clap 3.0 is no longer beta, and the method used to
generate the arguments was depreciated, so argument parsing had to be
rewritten to use Clap's builder system instead.

This also updates the program to build against Rust 2021 instead of
2018, however at this moment this change has no effect.